Corrected type of variable

+++ src/JsonTalker.cpp
@@ -87,7 +87,7 @@ const char* JsonTalker::_manifesto_name() const {
 
 
 void JsonTalker::_loop() {
-	if (_transmitted_message.active && millis() - _transmitted_message.identity > TALKIE_MAX_TTL) {
+	if (_transmitted_message.active && (uint16_t)millis() - _transmitted_message.identity > TALKIE_MAX_TTL) {
 		_transmitted_message.active = false;
 	}
 	if (_manifesto) _manifesto->_loop(*this);
